[PLAYER-5446] Move components which are used only in their panels
