+++ sdk/react/src/panels/AdPlaybackScreen/AdBar/AdBar.styles.json
