+++ sdk/react/src/panels/EndScreen/InfoPanel/InfoPanel.styles.json
