+++ sdk/react/src/panels/VideoView/UpNext/UpNext.styles.json
