+++ sdk/react/src/panels/VideoView/VideoWatermark/VideoWatermark.styles.json
